@@ -53,7 +53,7 @@ export class EventsClient extends BaseClient {
     super.apiCall({ method: 'post', path: '/events', body }, callback);
   }
 
-  getLatestEvents(channel: string, limit: number = 100, callback: (err: Error, events: []) => void) {
+  getLatestEvents(channel: string, limit = 100, callback: (err: Error, events: []) => void) {
     const qs = Object.assign({
       channel,
       limit,
